+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 This project uses [Semantic Versioning](https://semver.org).
 
 ## [Unreleased][unrelease]
+* Added support for dev environments, vcr and webmock as well configured rubocop and added a class to configure the gem itself
 * initial commit